Rename flash-loader to firmware
No more flash-loader/flash loader tool confusion...
