+++ .travis.yml
@@ -70,7 +70,7 @@ jobs:
         - python3 -m pip install 32blit bitstring construct
       after_success:
         - curl -O https://raw.githubusercontent.com/danvk/travis-weigh-in/adcc3a3e5bc16b360aa379d0fc23c3a6e1d9b21d/weigh_in.py
-        - python2 weigh_in.py firmware/flash-loader/flash-loader.bin
+        - python2 weigh_in.py firmware/firmware.bin
       before_deploy:
         - make install
         - tar -zcf ${RELEASE_FILE}.tar.gz bin/
