+++ docs/32Blit-Firmware.md
@@ -32,7 +32,7 @@ cmake .. -DCMAKE_TOOLCHAIN_FILE=../32blit.toolchain
 Now build the firmware binary and dfu files (for Linux/macOS and Windows flashing respectively):
 
 ```
-make flash-loader
+make firmware
 ```
 
 ## Flashing The Firmware To Your 32Blit
@@ -46,24 +46,24 @@ To enter DFU mode either hold the X & Y buttons and press the reset button or se
 Install `dfu-util` from your package manager then enter:
 
 ```
-sudo dfu-util -a 0 -s 0x08000000 -D firmware/flash-loader/flash-loader.bin
+sudo dfu-util -a 0 -s 0x08000000 -D firmware/firmware.bin
 ```
 
 ### Windows
 
-You will need [DfuSe Demonstration from st.com](https://www.st.com/en/development-tools/stsw-stm32080.html)) to flash `flash-loader.dfu` to your device.
+You will need [DfuSe Demonstration from st.com](https://www.st.com/en/development-tools/stsw-stm32080.html)) to flash `firmware.dfu` to your device.
 
 Once installed you should be able, from the `build.stm32` directory you created earlier, to flash 32Blit's firmware by running:
 
 ```
-../tools/wsl-flash firmware/flash-loader/flash-loader.dfu
+../tools/wsl-flash firmware/firmware.dfu
 ```
 
 Note: `wsl-flash` uses a hard-coded path to DfuSe 3.0.6 in `c:\Program Files (x86)`
 
 If this fails you can run the `DfusSeDemo` application and pick your 32Blit (it should be "STM Device in DFU Mode") from the "Available USB Devices" drop down.
 
-In the "Upload Action" section hit "Choose" and select the `flash-loader.dfu` file you built earlier. (It should be in `build.stm32/firmware/flash-loader/flash-loader.dfu`) and finally hit "Upgrade" to flash the file.
+In the "Upload Action" section hit "Choose" and select the `firmware.dfu` file you built earlier. (It should be in `build.stm32/firmware/firmware.dfu`) and finally hit "Upgrade" to flash the file.
 
 ## Troubleshooting
