+++ firmware/firmware.cpp
@@ -1,4 +1,4 @@
-#include "flash-loader.hpp"
+#include "firmware.hpp"
 #include "graphics/color.hpp"
 #include <cmath>
 #include "quadspi.h"
