--- firmware/flash-loader/CMakeLists.txt
@@ -1,11 +1,0 @@
-cmake_minimum_required(VERSION 3.8)
-project (flash-loader)
-include (../../32blit.cmake)
-
-blit_executable_int_flash(flash-loader flash-loader.cpp)
-
-
-target_compile_definitions(flash-loader
-  PRIVATE
- 	-DCDC_FIFO_BUFFERS=${CDC_FIFO_BUFFERS}
-)
